+++ src/main/java/com/taxicalls/passenger/repository/PassengerRepository.java
@@ -3,7 +3,7 @@ package com.taxicalls.passenger.repository;
 import com.taxicalls.passenger.model.Passenger;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PassengerRepository extends CrudRepository<Passenger, Integer> {
+public interface PassengerRepository extends CrudRepository<Passenger, Long> {
 
     public Passenger findByEmail(String email);
